README.md
This is a simple prototype for a Cassandra database controller. The code generates a text file of specified size which is inserted line-by-line to a newly created Cassandra table. Read through the comments in createDB.java for code description.

 **This code will not work under the following three conditions** 
- if Cassandra is not running on your system.
- if the keyspace "demo" does not exist.
- if the table the code tries to create already exists.

To run Cassandra on your system, ensure its latest version is installed and enter "Cassandra -f" if the terminal.

Use the CQL shell ("cqlsh" in terminal) to ensure that the keyspace "demo" is created ("describe keyspace demo;"), and that the table being created does not exist before running the code ("describe table demo.users;").

If demo does not exist, input "create keyspace demo with replication={'class':'SimpleStrategy', 'replication_factor':1};" into cqlsh. If the table being created exists already, input "drop table <table_name>;"